Code review followup
